requirements.txt
pillow~=10.0.0
Flask~=2.3.2

googletrans~=3.0.0
translate~=3.6.1
langdetect~=1.0.9



qrcode~=7.4.2


